Payload_Type/sliverserver/requirements.txt
mythic-container==0.4.18
tabulate
sliver-py
